+++ __tests__/unit/decorators/logger/logger.test.ts
@@ -1,19 +1,22 @@
 import { logContext, logInputParams } from '@decorators/logger/logger';
-import { Logger } from '@logger/logger';
-import { SampleClass } from '@tests/fixtures/class';
+import { LoggerImpl } from '@logger/logger';
+import { ILogger } from '@server/logger';
 
-jest.spyOn(Logger.prototype, 'setContext');
-jest.spyOn(Logger.prototype, 'debug');
+jest.spyOn(LoggerImpl.prototype, 'setContext');
+jest.spyOn(LoggerImpl.prototype, 'debug');
+
+@logContext()
+class SampleClass {
+  public logger: ILogger = new LoggerImpl();
+
+  @logInputParams()
+  public sampleMethod(_params: any): any {
+    this.logger.info('bla bla');
+  }
+}
 
 describe('decorators - logger', () => {
   it('should log param with caller context when apply logger decorators', () => {
-    logInputParams()(
-      SampleClass,
-      'sampleMethod',
-      Object.getOwnPropertyDescriptor(SampleClass.prototype, 'sampleMethod'),
-    );
-    logContext()(SampleClass);
-
     const sampleClass = new SampleClass();
     sampleClass.sampleMethod('test class decorator');
     expect(sampleClass.logger.setContext).toBeCalledWith('SampleClass', 'sampleMethod');